requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-dotenv==1.0.1
requests==2.32.3
jmespath==1.0.1
fastjsonschema==2.19.1
prometheus-client==0.20.0
pynacl==1.5.0
dnspython==2.6.1
idna==3.7
openai==0.28.0
